requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bandit==1.7.7
bcrypt==4.1.2
billiard==4.2.1
black==24.4.2
bleach==6.2.0
boto3==1.40.1
botocore==1.40.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
channels==4.1.0
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.9.2
cryptography==45.0.4
daphne==4.2.0
distlib==0.3.9
Django==5.2.2
django-appconf==1.1.0
django-extensions==4.1
django-formtools==2.5.1
django-model-utils==5.0.0
django-phonenumber-field==8.1.0
django-select2==8.4.1
django-simple-history==3.10.1
django-stubs==5.2.2
django-stubs-ext==5.2.2
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.1
drf-nested-routers==0.93.5
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==24.9.0
filelock==3.18.0
freezegun==1.5.3
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==5.10.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
pdfminer.six==20250506
pdfplumber==0.11.7
phonenumbers==9.0.7
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyOpenSSL==25.1.0
pyotp==2.9.0
pypdfium2==4.30.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.3
s3transfer==0.13.1
sentry-sdk==2.34.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stevedore==5.4.1
text-unidecode==1.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
twilio==9.2.3
urllib3==2.5.0
uvicorn==0.35.0
validate_docbr==1.11.1
vine==5.1.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
zope.interface==7.2
